test: refactor tests, extract repeating logic to helper funcs
